refactor job templates test to shared file
Issue: AAP-22947
